Rename package to zikade
