+++ routing/dht/zikade/options.go
@@ -1,4 +1,4 @@
-package dhtv2
+package zikade
 
 import (
 	"context"
